+++ app/controllers/notifications_controller.rb
@@ -53,6 +53,7 @@ class NotificationsController < ApplicationController
     return [@notification.project, @notification.adverse_event] if @notification.adverse_event
     return @notification.comment if @notification.comment
     return subject_event_redirect_path if @notification.handoff
+    return [@notification.project, @notification.sheet_unlock_request.sheet] if @notification.sheet_unlock_request
     notifications_path
   end