+++ db/migrate/20160316174017_add_auto_lock_sheets_to_projects.rb
@@ -1,0 +1,5 @@
+class AddAutoLockSheetsToProjects < ActiveRecord::Migration
+  def change
+    add_column :projects, :auto_lock_sheets, :string, null: false, default: 'never'
+  end
+end