+++ db/migrate/20160316190258_add_unlocked_at_to_sheets.rb
@@ -1,0 +1,5 @@
+class AddUnlockedAtToSheets < ActiveRecord::Migration
+  def change
+    add_column :sheets, :unlocked_at, :datetime
+  end
+end