+++ db/migrate/20160318155543_add_sheet_unlock_request_id_to_notifications.rb
@@ -1,0 +1,6 @@
+class AddSheetUnlockRequestIdToNotifications < ActiveRecord::Migration
+  def change
+    add_column :notifications, :sheet_unlock_request_id, :integer
+    add_index :notifications, :sheet_unlock_request_id
+  end
+end