+++ test/controllers/survey_controller_test.rb
@@ -168,10 +168,10 @@ class SurveyControllerTest < ActionController::TestCase
     assert_redirected_to about_survey_path(survey: assigns(:design).slug)
   end
 
-  test 'should not edit locked sheet survey' do
+  test 'should not edit auto-locked sheet survey' do
     get :edit, params: {
-      slug: designs(:admin_public_design).slug,
-      sheet_authentication_token: sheets(:external_locked).authentication_token
+      slug: designs(:auto_lock).slug,
+      sheet_authentication_token: sheets(:auto_lock).authentication_token
     }
     assert_not_nil assigns(:project)
     assert_not_nil assigns(:sheet)
@@ -211,6 +211,17 @@ class SurveyControllerTest < ActionController::TestCase
     assert_not_nil assigns(:project)
     assert_nil assigns(:sheet)
     assert_equal 'This survey no longer exists.', flash[:alert]
+    assert_redirected_to about_survey_path(survey: assigns(:design).slug)
+  end
+
+  test 'should not resubmit auto-locked sheet survey' do
+    patch :update, params: {
+      slug: designs(:auto_lock).slug,
+      sheet_authentication_token: sheets(:auto_lock).authentication_token
+    }
+    assert_not_nil assigns(:project)
+    assert_not_nil assigns(:sheet)
+    assert_equal 'This survey has been locked.', flash[:alert]
     assert_redirected_to about_survey_path(survey: assigns(:design).slug)
   end
 end